refactor: validation 테스트 분리 완료
